Use getScoutKey to access a model's primary key
When using binary ids in a project there is a problem when you want to flush all your records. In my project I override the getScoutKey method to return the text representation of the key. Now I run into the problem that this method isn't used constantly everywhere. So I propose to change the getter for the model's primary key from getKey to getScoutKey. The model needs to implement the Searchable trait anyway and moreover the default getScoutKey method in this trait refers to the getKey method.